pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "self-pruning-network"
version = "0.1.0"
description = "Interview-ready self-pruning neural network project with training, reporting, and FastAPI serving."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "matplotlib>=3.8.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pillow>=10.4.0",
    "pydantic>=2.9.0",
    "python-multipart>=0.0.9",
    "PyYAML>=6.0.2",
    "tabulate>=0.9.0",
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "tqdm>=4.66.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
